* release-6.1.2: (672 commits)
  fix trigger db
  Use caffeine for cache
  return totalpages = 0 for empty mappings
  clean up Dockerfile
  clean up Dockerfile
  clean up Dockerfile
  clean up Dockerfile
  change from runtime to implementation because of queries on Run information view.
  change CI, update templates from David
  change CI, remove openshift deployment because we will change it to helm
  add tzero data access
  add new plugin to get tzero information on calibration loop
  first working version with timeserie
  Do not remove locked tags
  fix role name
  add roles to specific paths
  add roles to specific paths
  add roles to specific paths
  change admin role
  Add database security for P1 usage
  ...

# Conflicts:
#	CREST_DOC.md
#	Dockerfile
#	README.md
#	USAGE.md
#	build.gradle
#	crestdb-client/curl/crest-test.sh
#	crestdb-client/curl/globaltag-methods.sh
#	crestdb-client/curl/tag-methods.sh
#	crestdb-client/curl/test-multi-upload.sh
#	crestdb-web/.swagger-codegen-ignore
#	crestdb-web/src/main/resources/application-default.yml
#	crestdb-web/src/test/resources/application-test.yml
#	entrypoint.sh
#	javaopts.properties
#	logback.xml.crest
#	mddocs/.openapi-generator-ignore
#	openapi/bundle/crestApi_all.yml
#	openapi/crestApi_all.yml
#	openapi/crestAtlasApi_all.yml
#	openapi/crestdb_atlasruninfoapi.yml
#	openapi/crestdb_globaltagsapi.yml
#	openapi/crestdb_runinfoapi.yml
#	openapi/crestdb_tagsapi.yml
#	openapi/crestdb_tagsmetaapi.yml
#	openapi/globaltags_api.yml
#	openapi/schemas/admin.yml
#	openapi/schemas/globaltagmaps.yml
#	openapi/schemas/globaltags.yml
#	openapi/schemas/iovs.yml
#	openapi/schemas/monitoring.yml
#	openapi/schemas/payloads.yml
#	openapi/schemas/runinfo.yml
#	openapi/schemas/tags.yml
#	openapi/schemas/tagsmeta.yml
#	openapi/tags_api.yml
#	openapi/templates/JavaJaxRS/cxf/server/openapi-generator-ignore.mustache
#	openapi/templates/JavaJaxRS/model.mustache
#	plugins/nats/java/plugin/nats/model/InputData.java
#	sql/crestdb.lal.psql
#	src/gen/java/hep/crest/server/swagger/api/GlobaltagsApi.java
#	src/gen/java/hep/crest/server/swagger/api/GlobaltagsApiService.java
#	src/gen/java/hep/crest/server/swagger/api/RuninfoApi.java
#	src/gen/java/hep/crest/server/swagger/api/TagsApi.java
#	src/gen/java/hep/crest/server/swagger/api/TagsApiService.java
#	src/gen/java/hep/crest/server/swagger/model/CrestBaseResponse.java
#	src/gen/java/hep/crest/server/swagger/model/FolderDto.java
#	src/gen/java/hep/crest/server/swagger/model/FolderSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/GenericMap.java
#	src/gen/java/hep/crest/server/swagger/model/GlobalTagDto.java
#	src/gen/java/hep/crest/server/swagger/model/GlobalTagMapDto.java
#	src/gen/java/hep/crest/server/swagger/model/GlobalTagMapSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/GlobalTagSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/HTTPResponse.java
#	src/gen/java/hep/crest/server/swagger/model/IovDto.java
#	src/gen/java/hep/crest/server/swagger/model/IovPayloadDto.java
#	src/gen/java/hep/crest/server/swagger/model/IovPayloadSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/IovSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/PayloadDto.java
#	src/gen/java/hep/crest/server/swagger/model/PayloadSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/PayloadTagInfoDto.java
#	src/gen/java/hep/crest/server/swagger/model/PayloadTagInfoSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/RespPage.java
#	src/gen/java/hep/crest/server/swagger/model/RunLumiInfoDto.java
#	src/gen/java/hep/crest/server/swagger/model/RunLumiSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/StoreDto.java
#	src/gen/java/hep/crest/server/swagger/model/StoreSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagMetaDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagMetaSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagSetDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagSummaryDto.java
#	src/gen/java/hep/crest/server/swagger/model/TagSummarySetDto.java
#	src/main/java/hep/crest/server/Application.java
#	src/main/java/hep/crest/server/aspects/IovSynchroAspect.java
#	src/main/java/hep/crest/server/aspects/TagSecurityAspect.java
#	src/main/java/hep/crest/server/aspects/UserInfoImpl.java
#	src/main/java/hep/crest/server/config/CacheConfig.java
#	src/main/java/hep/crest/server/config/filters/CORSFilter.java
#	src/main/java/hep/crest/server/controllers/JsonStreamProcessor.java
#	src/main/java/hep/crest/server/converters/CustomMapper.java
#	src/main/java/hep/crest/server/converters/GlobalTagMapMapper.java
#	src/main/java/hep/crest/server/converters/TagMapper.java
#	src/main/java/hep/crest/server/data/pojo/CrestRoles.java
#	src/main/java/hep/crest/server/data/pojo/GlobalTag.java
#	src/main/java/hep/crest/server/data/pojo/Tag.java
#	src/main/java/hep/crest/server/repositories/triggerdb/TriggerDb.java
#	src/main/java/hep/crest/server/security/SecurityConfig.java
#	src/main/java/hep/crest/server/security/SecurityDefaultConfig.java
#	src/main/java/hep/crest/server/services/CachePayloadBuffer.java
#	src/main/java/hep/crest/server/services/FolderService.java
#	src/main/java/hep/crest/server/services/GlobalTagMapService.java
#	src/main/java/hep/crest/server/services/GlobalTagService.java
#	src/main/java/hep/crest/server/services/IovService.java
#	src/main/java/hep/crest/server/services/PayloadService.java
#	src/main/java/hep/crest/server/services/TagMetaService.java
#	src/main/java/hep/crest/server/services/TagService.java
#	src/main/java/hep/crest/server/swagger/impl/AdminApiServiceImpl.java
#	src/main/java/hep/crest/server/swagger/impl/FoldersApiServiceImpl.java
#	src/main/java/hep/crest/server/swagger/impl/GlobaltagmapsApiServiceImpl.java
#	src/main/java/hep/crest/server/swagger/impl/GlobaltagsApiServiceImpl.java
#	src/main/java/hep/crest/server/swagger/impl/PayloadsApiServiceImpl.java
#	src/main/java/hep/crest/server/swagger/impl/TagsApiServiceImpl.java
#	src/main/resources/application-default.yml
#	src/main/resources/application-defaultredis.yml
#	src/main/resources/application-keycloak.yml
#	src/main/resources/application.yml.orig
#	src/main/resources/logback.xml
#	src/test/java/hep/crest/server/MapperTest.java
#	src/test/java/hep/crest/server/TestCrestMappings.java
#	src/test/java/hep/crest/server/TestCrestRunLumi.java
#	src/test/java/hep/crest/server/TestCrestTags.java
#	src/test/java/hep/crest/server/TestRedisIntegration.java
#	src/test/java/hep/crest/server/TestSecurityConfig.java
#	src/test/resources/application-test.yml
#	src/test/resources/application-testredis.yml

+++ src/main/java/hep/crest/server/services/GlobalTagMapService.java
@@ -5,6 +5,7 @@ package hep.crest.server.services;
 
 import hep.crest.server.data.pojo.GlobalTag;
 import hep.crest.server.data.pojo.GlobalTagMap;
+import hep.crest.server.data.pojo.GlobalTagTypeEnum;
 import hep.crest.server.data.pojo.Tag;
 import hep.crest.server.data.repositories.GlobalTagMapRepository;
 import hep.crest.server.data.repositories.GlobalTagRepository;
@@ -12,6 +13,7 @@ import hep.crest.server.data.repositories.TagRepository;
 import hep.crest.server.exceptions.AbstractCdbServiceException;
 import hep.crest.server.exceptions.CdbNotFoundException;
 import hep.crest.server.exceptions.ConflictException;
+import hep.crest.server.swagger.model.TagDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -41,20 +43,25 @@ public class GlobalTagMapService {
      * Repository.
      */
     private TagRepository tagRepository;
+    /**
+     * Service.
+     */
+    private TagService tagService;
 
     /**
      * Ctor with injection.
      * @param globalTagMapRepository
      * @param globalTagRepository
-     * @param tagRepository
+     * @param tagService
      */
     @Autowired
     public GlobalTagMapService(GlobalTagMapRepository globalTagMapRepository,
                                GlobalTagRepository globalTagRepository,
-                               TagRepository tagRepository) {
+                               TagService tagService) {
         this.globalTagMapRepository = globalTagMapRepository;
         this.globalTagRepository = globalTagRepository;
-        this.tagRepository = tagRepository;
+        this.tagService = tagService;
+        this.tagRepository = tagService.getTagRepository();
     }
 
 
@@ -62,7 +69,7 @@ public class GlobalTagMapService {
      * @param gtName the String represnting the GlobalTag name
      * @return Iterable<GlobalTagMap>
      */
-    public Iterable<GlobalTagMap> getTagMap(String gtName) {
+    public List<GlobalTagMap> getTagMap(String gtName) {
         log.debug("Search for GlobalTagMap entries by GlobalTag name {}", gtName);
         return globalTagMapRepository.findByGlobalTagName(gtName);
     }
@@ -71,7 +78,7 @@ public class GlobalTagMapService {
      * @param tagName the String
      * @return Iterable<GlobalTagMapDto>
      */
-    public Iterable<GlobalTagMap> getTagMapByTagName(String tagName) {
+    public List<GlobalTagMap> getTagMapByTagName(String tagName) {
         log.debug("Search for GlobalTagMap entries by Tag name {}", tagName);
         return globalTagMapRepository.findByTagName(tagName);
     }
@@ -117,13 +124,13 @@ public class GlobalTagMapService {
      * @param mrecord   the record
      * @return the iterable
      */
-    public Iterable<GlobalTagMap> findMapsByGlobalTagLabelTag(String globaltag, String label,
+    public List<GlobalTagMap> findMapsByGlobalTagLabelTag(String globaltag, String label,
                                                               String tag,
                                                               String mrecord) {
         log.debug("Search for GlobalTagMap entries by Global, Label, Tag and eventually filter by"
                   + " record : {} {} {} {}",
                 globaltag, label, tag, mrecord);
-        if (tag == null || tag.isEmpty()) {
+        if (tag == null || tag.isEmpty() || tag.equals("none")) {
             tag = "%";
         }
         tag = tag.replace("*", "%");
@@ -145,7 +152,7 @@ public class GlobalTagMapService {
      * @return the global tag map
      */
     @Transactional
-    public GlobalTagMap deleteMap(GlobalTagMap entity) {
+    protected GlobalTagMap deleteMap(GlobalTagMap entity) {
         globalTagMapRepository.delete(entity);
         return entity;
     }
@@ -160,11 +167,87 @@ public class GlobalTagMapService {
     public List<GlobalTagMap> deleteMapList(Iterable<GlobalTagMap> entitylist) {
         List<GlobalTagMap> deletedlist = new ArrayList<>();
         for (GlobalTagMap globalTagMap : entitylist) {
+            GlobalTag globalTag = globalTagMap.getGlobalTag();
+            if (globalTag.getType() == GlobalTagTypeEnum.LOCKED.getCode()) {
+                log.warn("GlobalTag {} is locked, cannot delete mapping {}",
+                        globalTag.getName(), globalTagMap);
+                throw new ConflictException("GlobalTag is locked, cannot delete mapping");
+            }
+            Tag tag = globalTagMap.getTag();
+            if (tag.getStatus().equals(TagDto.StatusEnum.LOCKED.toString())) {
+                log.warn("Tag {} is locked, cannot delete mapping {}",
+                        tag.getName(), globalTagMap);
+                throw new ConflictException("Tag is locked, cannot delete mapping");
+            }
+            log.info("Delete GlobalTagMap entry {}: {} {}", globalTagMap, globalTag, tag);
             GlobalTagMap delentity = this.deleteMap(globalTagMap);
             if (delentity != null) {
                 deletedlist.add(delentity);
             }
         }
         return deletedlist;
+    }
+
+    /**
+     * Check the type of the tag, using the list of GlobalTagMap entries.
+     * If a locked global tag is found, the type is locked.
+     * When locked, a tag takes by default a synchronized status of append-only, in case
+     * the synchronization is not set (or set to none).
+     *
+     * @param globaltag the GlobalTag to check
+     */
+    @Transactional
+    public void updateAssociatedTagsType(GlobalTag globaltag) {
+        log.debug("Check tag type for list of GlobalTagMap entries");
+        List<GlobalTagMap> mapList = this.getTagMap(globaltag.getName());
+        // Check if the list is empty.
+        if (mapList.isEmpty()) {
+            log.info("No GlobalTagMap entries found.");
+            return;
+        }
+        char gtype = globaltag.getType();
+        for (GlobalTagMap map : mapList) {
+            if (GlobalTagTypeEnum.fromCode(gtype) == GlobalTagTypeEnum.LOCKED) {
+                Tag tag = map.getTag();
+                tag.setStatus(TagDto.StatusEnum.LOCKED.toString());
+                String sync = tag.getSynchronization();
+                if (sync == null || sync.isEmpty() || sync.equals("ALL")) {
+                    log.warn("Tag {} is locked, setting synchronization to append-only",
+                            tag.getName());
+                    tag.setSynchronization(TagDto.SynchronizationEnum.SV.toString());
+                }
+                tagService.updateTag(tag);
+            }
+            else {
+                Tag tag = updateTagStatus(map.getTag().getName(),
+                        TagDto.StatusEnum.UNLOCKED.toString());
+                log.debug("Tag status after attempt to unlock : {} ", tag);
+            }
+        }
+    }
+
+    /**
+     * @param tagname the String
+     * @param status  the String
+     * @return Tag
+     * @throws AbstractCdbServiceException If an Exception occurred
+     */
+    public Tag updateTagStatus(String tagname, String status) throws AbstractCdbServiceException {
+        log.debug("Update Tag status for {} to {}", tagname, status);
+        List<GlobalTagMap> tagMaps = this.getTagMapByTagName(tagname);
+        Tag tag = tagRepository.findById(tagname).orElseThrow(
+                () -> new CdbNotFoundException("Tag does not exists for name " + tagname));
+        for (GlobalTagMap map : tagMaps) {
+            char gtypechar = map.getGlobalTag().getType();
+            if (GlobalTagTypeEnum.fromCode(gtypechar) == GlobalTagTypeEnum.LOCKED) {
+                log.warn("Tag {} is associated to a locked global tag {}, cannot change status to"
+                                + " {}",
+                        tagname, map.getGlobalTag().getName(), status);
+                throw new ConflictException("Tag is locked, cannot change status to " + status);
+            }
+        }
+        // No constraints found, update the tag.
+        tag.setStatus(TagDto.StatusEnum.valueOf(status).toString());
+        return tagService.updateTag(tag);
     }
 }
